SCIM Connector/requirements.txt
Flask==3.0.0
Flask-HTTPAuth==4.8.0
pyodbc==5.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1